requirements.txt
geopandas
shapely
pyproj
matplotlib
folium
pandas
requests
networkx
pyyaml